+++ CHANGELOG.md
@@ -4,10 +4,10 @@
 
 - Update Ruby dependency to >= 3.0.5
 - Handle simple form-data in request bodies (see https://github.com/ahx/openapi_first/issues/149)
-- Update to hanami-router 2.0.0 stable. Use it's BodyParser middleware to parse the body.
-- Support simple form data requests. Still have to figure out file uploads.
+- Update to hanami-router 2.0.0 stable
 
 ## 0.20.0
+
 - You can pass a filepath to `spec:` now so you no longer have to call `OpenapiFirst.load` anymore.
 - Router is optional now.
   You no longer have to add `Router` to your middleware stack. You still can add it to customize behaviour by setting options, but you no longer have to add it.